+++ packages/grafana-ui/src/components/ContextMenu/ContextMenu.story.tsx
@@ -4,12 +4,11 @@ import React from 'react';
 
 import { withCenteredStory } from '../../utils/storybook/withCenteredStory';
 import { IconButton } from '../IconButton/IconButton';
-import { MenuGroup } from '../Menu/MenuGroup';
-import { MenuItem } from '../Menu/MenuItem';
 
 import { ContextMenu, ContextMenuProps } from './ContextMenu';
 import mdx from './ContextMenu.mdx';
-import { WithContextMenu } from './WithContextMenu';
+import { renderMenuItems } from './ContextMenuStoryHelper';
+import { WithContextMenu, WithContextMenuProps } from './WithContextMenu';
 
 const meta: ComponentMeta<typeof ContextMenu> = {
   title: 'General/ContextMenu',
@@ -20,37 +19,15 @@ const meta: ComponentMeta<typeof ContextMenu> = {
       page: mdx,
     },
     controls: {
-      exclude: ['renderMenuItems', 'renderHeader'],
+      exclude: ['renderMenuItems', 'renderHeader', 'onClose', 'children'],
     },
   },
   args: {
     x: 200,
     y: 300,
     focusOnOpen: true,
+    renderMenuItems: renderMenuItems,
   },
-};
-
-const menuItems = [
-  {
-    label: 'Test',
-    items: [
-      { label: 'First', ariaLabel: 'First' },
-      { label: 'Second', ariaLabel: 'Second' },
-      { label: 'Third', ariaLabel: 'Third' },
-      { label: 'Fourth', ariaLabel: 'Fourth' },
-      { label: 'Fifth', ariaLabel: 'Fifth' },
-    ],
-  },
-];
-
-const renderMenuItems = () => {
-  return menuItems.map((group, index) => (
-    <MenuGroup key={`${group.label}${index}`} label={group.label}>
-      {group.items.map((item) => (
-        <MenuItem key={item.label} label={item.label} />
-      ))}
-    </MenuGroup>
-  ));
 };
 
 const renderHeader = (): React.ReactNode => {
@@ -68,9 +45,9 @@ export const Basic: ComponentStory<typeof ContextMenu> = (args: ContextMenuProps
   );
 };
 
-export const WithState = () => {
+export const WithState: ComponentStory<typeof WithContextMenu> = (args: WithContextMenuProps) => {
   return (
-    <WithContextMenu renderMenuItems={renderMenuItems}>
+    <WithContextMenu {...args}>
       {({ openMenu }) => <IconButton name="info-circle" onClick={openMenu} />}
     </WithContextMenu>
   );
